# httpServer/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/apiGroups": {
            "get": {
                "tags": [
                    "ApiGroup"
                ],
                "summary": "List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query, ?query=age:\u003e:21,name:like:%jason%",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "order, ?order=age:desc,created_at:asc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "apiGroup array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ApiGroup"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ApiGroup"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "description": "ApiGroup",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ApiGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/apiGroups/{apiGroupId}": {
            "get": {
                "tags": [
                    "ApiGroup"
                ],
                "summary": "Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "apiGroupId",
                        "name": "apiGroupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "apiGroup object",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ApiGroup"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ApiGroup"
                ],
                "summary": "Put",
                "parameters": [
                    {
                        "description": "apiGroup",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ApiGroup"
                        }
                    },
                    {
                        "type": "string",
                        "description": "apiGroupId",
                        "name": "apiGroupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ApiGroup"
                ],
                "summary": "Delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "apiGroupId",
                        "name": "apiGroupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "ApiGroup"
                ],
                "summary": "Patch",
                "parameters": [
                    {
                        "description": "apiGroup",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ApiGroup"
                        }
                    },
                    {
                        "type": "string",
                        "description": "apiGroupId",
                        "name": "apiGroupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/apis": {
            "get": {
                "tags": [
                    "Api"
                ],
                "summary": "List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query, ?query=age:\u003e:21,name:like:%jason%",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "order, ?order=age:desc,created_at:asc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "api array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Api"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Api"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "description": "Api",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.Api"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/apis/{apiId}": {
            "get": {
                "tags": [
                    "Api"
                ],
                "summary": "Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "apiId",
                        "name": "apiId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "api object",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.Api"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Api"
                ],
                "summary": "Put",
                "parameters": [
                    {
                        "description": "api",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.Api"
                        }
                    },
                    {
                        "type": "string",
                        "description": "apiId",
                        "name": "apiId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Api"
                ],
                "summary": "Delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "apiId",
                        "name": "apiId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Api"
                ],
                "summary": "Patch",
                "parameters": [
                    {
                        "description": "api",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.Api"
                        }
                    },
                    {
                        "type": "string",
                        "description": "apiId",
                        "name": "apiId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Api": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "describe": {
                    "type": "string"
                },
                "groupId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.ApiGroup": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "describe": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}